% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facor.R
\name{facor}
\alias{facor}
\title{Principle analysis and correlation test result}
\usage{
facor(
  data,
  lif = c("lif1", "lif2", "lif3"),
  peo = c("peo1g", "peo2p", "peo3g"),
  gen = c("gen1", "gen2", "gen3"),
  per = c("per1", "per2", "per3"),
  ev = c("gen", "peo", "lif", "per")
)
}
\arguments{
\item{data}{A data frame used to compare}

\item{lif}{A vector of variable names for latent variable "life tolerance"}

\item{peo}{A vector of variable names for latent variable "people's voice"}

\item{gen}{A vector of variable names for latent variable "gender equality"}

\item{per}{A vector of variable names for latent variable "personal autonomy"}

\item{ev}{A vector of variable names for latent variable "emancipative values"}
}
\value{
A list of two components. The first is a data frame used to create correlation plot. The second is a list consist of results of principle components analysis, including KMO, Cronbach's alpha and loadings
}
\description{
Principle analysis and correlation test result
}
