% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misintt.R
\name{misintt}
\alias{misintt}
\title{Missing data totally in some groups calculation.}
\usage{
misintt(data, varnames, group)
}
\arguments{
\item{data}{A data frame with missing values}

\item{varnames}{A vector of characters consists of names of those variables that need to calculate.}

\item{group}{A character of group variable name.}
}
\value{
a data frame with missing value proportion.
}
\description{
Missing data totally in some groups calculation.
}
