% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoding.R
\name{recoding}
\alias{recoding}
\title{Recoding variables using Welzel(2011)'s methods in data}
\usage{
recoding(data)
}
\arguments{
\item{data}{A dataframe including variables names "gen1","gen2","gen3","peo1","peo2","peo3","peo4","lif1","lif2","lif3"}
}
\value{
A dataframe including the recoded variables above
}
\description{
Recoding variables using Welzel(2011)'s methods in data
}
