% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Random effects summary of simulations}
\usage{
simdata(
  simfix,
  simran,
  varfix,
  varran,
  interval,
  interaction = F,
  movarfix,
  movarran,
  factorname = c("non-Confucian Societies", "Confucian Societies")
)
}
\arguments{
\item{simfix}{A data frame created by arm::sim(), the fixed effect in the result.}

\item{simran}{A data frame created by arm::sim(), the random effect in the result.}

\item{varfix}{A character of a variable name used in the fixed effect.}

\item{varran}{A vector of a set of variable names used in the random effect.}

\item{interval}{A number of confidence interval value.}

\item{interaction}{A logical value to confirm whether there exists a interaction. Default is False.}

\item{movarfix}{A character of a variable name used in the fixed effect containing modifier variable name.}

\item{movarran}{A character of a variable name used in the random effect containing modifier variable name.}

\item{factorname}{A vector of factor's name as for Chinese-English transformation. Default is c("non-Confucian Societies","Confucian Societies").}
}
\value{
A data frame
}
\description{
Random effects summary of simulations
}
